# Conflicts:
#	build.gradle
#	src/main/java/com/sparta/goods/domain/product/repository/ProductRepository.java

+++ src/main/java/com/sparta/goods/domain/product/entity/Product.java
@@ -22,7 +22,7 @@ public class Product {
 
     @Column
     @Min(100)
-    private Integer price;
+    private Double price;
 
     @Column
     @Min(0)
@@ -38,7 +38,7 @@ public class Product {
 
 
     @Builder
-    public Product(String productName, Integer price, Integer quantity, String introduction, GoodsCategory category) {
+    public Product(String productName, Double price, Integer quantity, String introduction, GoodsCategory category) {
         this.productName = productName;
         this.price = price;
         this.quantity = quantity;
